requirements.txt
antlr4-python3-runtime==4.8
appdirs==1.4.4
astroid==2.5.6
attackcti==0.3.4.4
attrs==21.2.0
beautifulsoup4==4.9.3
black==21.5b2
bs4==0.0.1
certifi==2021.5.30
charset-normalizer==2.0.2
click==8.0.1
cloudpickle==1.6.0
colorama==0.4.4
cssselect==1.1.0
Cython==0.29.24
dask==2021.7.0
dill==0.3.4
distributed==2021.7.0
entrypoints==0.3
feedfinder2==0.0.4
feedparser==6.0.8
filelock==3.0.12
flake8==3.9.2
Flask==2.0.1
fsspec==2021.7.0
HeapDict==1.0.1
html2text==2020.1.16
idna==3.2
isort==5.8.0
itsdangerous==2.0.1
jieba3k==0.35.1
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
lazy-object-proxy==1.6.0
locket==0.2.1
lxml==4.6.3
MarkupSafe==2.0.1
mccabe==0.6.1
msgpack==1.0.2
mypy-extensions==0.4.3
newspaper3k==0.2.8
nltk==3.6.2
numpy==1.21.0
pandas==1.3.0
partd==1.2.0
pathspec==0.8.1
Pillow==8.3.1
psutil==5.8.0
pyarrow==4.0.1
pycodestyle==2.7.0
pyflakes==2.3.1
pylint==2.8.2
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
regex==2021.4.4
requests==2.26.0
requests-file==1.5.1
scikit-learn==0.22.2.post1
scipy==1.7.0
sgmllib3k==1.0.0
simplejson==3.17.3
six==1.16.0
sklearn==0.0
sortedcontainers==2.4.0
soupsieve==2.2.1
stix2==3.0.0
stix2-patterns==1.3.2
taxii2-client==2.3.0
tblib==1.7.0
threadpoolctl==2.2.0
tika==1.24
tinysegmenter==0.3
tldextract==3.1.0
toml==0.10.2
toolz==0.11.1
tornado==6.1
tqdm==4.61.2
urllib3==1.26.6
Werkzeug==2.0.1
wrapt==1.12.1
zict==2.0.0
